Missing lines
git-svn-id: http://svn.l2jfree.com/l2jfree@6375 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/instancemanager/games/Lottery.java
@@ -506,7 +506,7 @@ public class Lottery
 
 	public long[] checkTicket(int id, int enchant, int type2)
 	{
-		int res[] =
+		long res[] =
 		{ 0, 0 };
 
 		Connection con = null;
